+++ docs/release-notes.rst
@@ -7,6 +7,8 @@ History
 
 * Generalized ``diagnostics()``, simplified internals for ``cli`` fixture
 
+* Added ``wcswidth()``
+
 
 2.7.5 (2021-06-04)
 -----------------